:wrench: Update: ESLint + tsconfig + prettier minor rules update
