+++ .prettierrc
@@ -1,5 +1,8 @@
 {
   "singleQuote": true,
   "semi": false,
-  "bracketSameLine": true
+  "bracketSameLine": true,
+  "trailingComma": "all",
+  "tabWidth": 1,
+  "printWidth": 100
 }
