+++ components/AnimatedLayer.client.tsx
@@ -13,38 +13,38 @@ type Props = {
 }
 
 const AnimatedLayer = (props: Props) => {
-  const {
-    speed,
-    src,
-    alt,
-    scrollY,
-    bg = false,
-    screenWidth,
-    customStyle = {},
-  } = props
+	const {
+		speed,
+		src,
+		alt,
+		scrollY,
+		bg = false,
+		screenWidth,
+		customStyle = {},
+	} = props
 
-  const inputRange = [0, 400, 500, 700]
-  const outputRange = bg ? inputRange : [0, 400, 550, 600]
-  const applySpeed = (array: number[]) => array.map((x) => x * speed)
+	const inputRange = [0, 400, 500, 700]
+	const outputRange = bg ? inputRange : [0, 400, 550, 600]
+	const applySpeed = (array: number[]) => array.map((x) => x * speed)
 
-  const y = useTransform(scrollY, inputRange, applySpeed(outputRange))
+	const y = useTransform(scrollY, inputRange, applySpeed(outputRange))
 
-  const style = { ...customStyle, y }
+	const style = { ...customStyle, y }
 
-  if (screenWidth >= 2560) {
-    return (
-      <figure className="z-1 absolute w-full">
-        <Image layout="responsive" src={src} alt={alt} />
-      </figure>
-    )
-  }
+	if (screenWidth >= 2560) {
+		return (
+			<figure className="z-1 absolute w-full">
+				<Image layout="responsive" src={src} alt={alt} />
+			</figure>
+		)
+	}
 
-  return (
-    <motion.figure style={style} className="z-1 absolute w-full">
-      <Image layout="responsive" src={src} alt={alt} />
-      <div className="relative z-10 h-[350px] w-full bg-main" />
-    </motion.figure>
-  )
+	return (
+		<motion.figure style={style} className="z-1 absolute w-full">
+			<Image layout="responsive" src={src} alt={alt} />
+			<div className="relative z-10 h-[350px] w-full bg-main" />
+		</motion.figure>
+	)
 }
 
 export default AnimatedLayer
