+++ components/Content.server.tsx
@@ -5,11 +5,11 @@ type Props = {
 }
 
 const Content = ({ children }: Props) => (
-  <div className="flex w-full justify-center">
-    <section id="content" className="w-full max-w-[1920px] px-[10rem]">
-      {children}
-    </section>
-  </div>
+	<div className="flex w-full justify-center">
+		<section id="content" className="w-full max-w-[1920px] px-[10rem]">
+			{children}
+		</section>
+	</div>
 )
 
 export default Content
