+++ components/Head.client.tsx
@@ -5,10 +5,10 @@ type Props = {
 }
 
 const Head = ({ title }: Props) => (
-  <NextHead>
-    <title>{title}</title>
-    <meta name="viewport" content="initial-scale=1.0, width=device-width" />
-  </NextHead>
+	<NextHead>
+		<title>{title}</title>
+		<meta name="viewport" content="initial-scale=1.0, width=device-width" />
+	</NextHead>
 )
 
 export default Head
