+++ components/Page.server.tsx
@@ -10,16 +10,16 @@ type Props = {
 }
 
 const Page = ({ children, header, title }: Props) => (
-  <Fragment>
-    {/* Page head meta */}
-    <Head title={title} />
-    {/* Navbar */}
-    <Navbar />
-    {/* Header if provided */}
-    {header && <header className="bg-main">{header}</header>}
-    {/* We intentionally leave out width restrictions for styling reasons (i.e. we want fluent 4k effects) */}
-    <main className="min-h-full w-full bg-main">{children}</main>
-  </Fragment>
+	<Fragment>
+		{/* Page head meta */}
+		<Head title={title} />
+		{/* Navbar */}
+		<Navbar />
+		{/* Header if provided */}
+		{header && <header className="bg-main">{header}</header>}
+		{/* We intentionally leave out width restrictions for styling reasons (i.e. we want fluent 4k effects) */}
+		<main className="min-h-full w-full bg-main">{children}</main>
+	</Fragment>
 )
 
 export default Page
