+++ package.json
@@ -6,6 +6,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
+    "lint:fix": "next lint --fix",
     "buildcss": "npx tailwindcss -o build.css --minify",
     "ci": "npm install --immutable --immutable-cache --check-cache"
   },
@@ -23,7 +24,7 @@
     "autoprefixer": "10.4.2",
     "eslint": "8.8.0",
     "eslint-config-next": "12.0.8",
-    "eslint-config-prettier": "8.3.0",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "7.28.0",
     "eslint-plugin-react-hooks": "4.3.0",
     "postcss": "8.4.5",
